refactor(Movie): 로직 도메인 로직으로 이동
